# More efficient membership check in AbstractZonotope