The previous implementation added a dummy dimension for optimization. But here we are only interested in feasibility, and zonotopes are always bounded, so that is redundant.
(This also allows to replace some vectors by single instances. They are internally converted to a vector again, but maybe one day we have a more efficient solver bridge.)


```julia
# master
julia> @time [1.0, 0.2] ∈ Z
  0.000699 seconds (1.80 k allocations: 65.766 KiB)
false

julia> @time [1.0, 0.1] ∈ Z
  0.000720 seconds (1.80 k allocations: 65.500 KiB)
true

# this PR:
julia> @time [1.0, 0.2] ∈ Z
  0.000492 seconds (1.70 k allocations: 59.750 KiB)
false

julia> @time [1.0, 0.1] ∈ Z
  0.000697 seconds (1.70 k allocations: 59.500 KiB)
true
```